+++ src/components/core/layout/index.ts
@@ -1,3 +1,4 @@
 export { default as Main } from './Main';
+export { default as Section } from './Section';
 export { default as Theme } from './Theme';
 export { default as Visibility } from './Visibility';